+++ chart/falco/tests/unit/containerPlugin/containerPluginConfig_test.go
@@ -11,7 +11,7 @@ import (
 	"gopkg.in/yaml.v3"
 	corev1 "k8s.io/api/core/v1"
 
-	"github.com/falcosecurity/charts/charts/falco/tests/unit"
+	"github.com/falcosecurity/falco/chart/falco/tests/unit"
 	"github.com/gruntwork-io/terratest/modules/helm"
 )