+++ chart/falco/tests/unit/falcoTemplates/falcoctlConfigVolume_test.go
@@ -1,0 +1,74 @@
+// SPDX-License-Identifier: Apache-2.0
+// Copyright 2026 The Falco Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package falcoTemplates
+
+import (
+	"path/filepath"
+	"testing"
+
+	"github.com/falcosecurity/falco/chart/falco/tests/unit"
+	"github.com/gruntwork-io/terratest/modules/helm"
+	"github.com/stretchr/testify/require"
+	appsv1 "k8s.io/api/apps/v1"
+	corev1 "k8s.io/api/core/v1"
+)
+
+func TestFalcoctlConfigVolume(t *testing.T) {
+	t.Parallel()
+
+	helmChartPath, err := filepath.Abs(unit.ChartPath)
+	require.NoError(t, err)
+
+	testCases := []struct {
+		name     string
+		values   map[string]string
+		expected func(t *testing.T, volumes []corev1.Volume)
+	}{
+		{
+			"defaultValues",
+			nil,
+			func(t *testing.T, volumes []corev1.Volume) {
+				require.NotNil(t, findVolume("falcoctl-config-volume", volumes))
+			},
+		},
+		{
+			"falcoctlArtifactsDisabled",
+			map[string]string{
+				"falcoctl.artifact.install.enabled": "false",
+				"falcoctl.artifact.follow.enabled":  "false",
+			},
+			func(t *testing.T, volumes []corev1.Volume) {
+				require.Nil(t, findVolume("falcoctl-config-volume", volumes))
+			},
+		},
+	}
+
+	for _, testCase := range testCases {
+		testCase := testCase
+
+		t.Run(testCase.name, func(t *testing.T) {
+			t.Parallel()
+
+			options := &helm.Options{SetValues: testCase.values}
+			output := helm.RenderTemplate(t, options, helmChartPath, unit.ReleaseName, []string{"templates/daemonset.yaml"})
+
+			var ds appsv1.DaemonSet
+			helm.UnmarshalK8SYaml(t, output, &ds)
+
+			testCase.expected(t, ds.Spec.Template.Spec.Volumes)
+		})
+	}
+}